+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.1.1 - Unreleased
 
+- Prevent late call-log and tool callbacks after teardown from crashing the Vox server. Thanks @joeVenner.
+
 ## 0.1.0 - 2026-06-11
 
 ### Added